+++ changelog.md
@@ -54,6 +54,7 @@
 - Renamed filter "eml_import_url_before" to "eml_import_url"
 - Renamed filter "eml_blacklist" to "eml_prevent_import"
 - Hosting of files can now only be changed by users with the capability to upload external files
+- Synced files will be linked with its linked source in media library
 
 ### Fixed